Change base namespace to CrEOF\Geo\WKT to avoid class collision with other CrEOF packages.
