+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file using the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0] 2015-11-18
+### Added
+- Change base namespace to CrEOF\Geo\WKT to avoid class collision with other CrEOF packages.
+
 ## [1.0.0] 2015-11-11
 ### Added
 - Initial release.
