+++ composer.json
@@ -21,7 +21,7 @@
     },
     "autoload": {
         "psr-0": {
-            "CrEOF\\WKT": "lib/"
+            "CrEOF\\Geo\\WKT": "lib/"
         }
     }
 }
