+++ lib/CrEOF/Geo/WKT/Exception/ExceptionInterface.php
@@ -21,15 +21,15 @@
  * SOFTWARE.
  */
 
-namespace CrEOF\WKT\Exception;
+namespace CrEOF\Geo\WKT\Exception;
 
 /**
- * UnexpectedValueException
+ * Exception interface for library exceptions
  *
  * @author  Derek J. Lambert <dlambert@dereklambert.com>
  * @license http://dlambert.mit-license.org MIT
  */
-class UnexpectedValueException extends \UnexpectedValueException implements ExceptionInterface
+interface ExceptionInterface
 {
 
 }
