+++ tests/CrEOF/Geo/WKT/Tests/LineStringParserTest.php
@@ -21,30 +21,28 @@
  * SOFTWARE.
  */
 
-namespace CrEOF\WKT\Tests;
+namespace CrEOF\Geo\WKT\Tests;
 
-use CrEOF\WKT\Parser;
+use CrEOF\Geo\WKT\Parser;
 
 /**
- * MULTIPOINT parser tests
+ * LINESTRING parser tests
  *
  * @author  Derek J. Lambert <dlambert@dereklambert.com>
  * @license http://dlambert.mit-license.org MIT
  */
-class MultiPointParserTest extends \PHPUnit_Framework_TestCase
+class LineStringParserTest extends \PHPUnit_Framework_TestCase
 {
-    public function testParsingMultiPointValue()
+    public function testParsingLineStringValue()
     {
-        $value    = 'MULTIPOINT(0 0,10 0,10 10,0 10)';
+        $value    = 'LINESTRING(34.23 -87, 45.3 -92)';
         $parser   = new Parser($value);
         $expected = array(
             'srid'  => null,
-            'type'  => 'MULTIPOINT',
+            'type'  => 'LINESTRING',
             'value' => array(
-                array(0, 0),
-                array(10, 0),
-                array(10, 10),
-                array(0, 10)
+                array(34.23, -87),
+                array(45.3, -92)
             )
         );
 
@@ -53,18 +51,16 @@ class MultiPointParserTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals($expected, $actual);
     }
 
-    public function testParsingMultiPointValueWithSrid()
+    public function testParsingLineStringValueWithSrid()
     {
-        $value    = 'SRID=4326;MULTIPOINT(0 0,10 0,10 10,0 10)';
+        $value    = 'SRID=4326;LINESTRING(34.23 -87, 45.3 -92)';
         $parser   = new Parser($value);
         $expected = array(
             'srid'  => 4326,
-            'type'  => 'MULTIPOINT',
+            'type'  => 'LINESTRING',
             'value' => array(
-                array(0, 0),
-                array(10, 0),
-                array(10, 10),
-                array(0, 10)
+                array(34.23, -87),
+                array(45.3, -92)
             )
         );
 
@@ -74,12 +70,24 @@ class MultiPointParserTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
-     * @expectedException        \CrEOF\WKT\Exception\UnexpectedValueException
-     * @expectedExceptionMessage [Syntax Error] line 0, col 11: Error: Expected CrEOF\WKT\Lexer::T_INTEGER, got "(" in value "MULTIPOINT((0 0,10 0,10 10,0 10))"
+     * @expectedException        \CrEOF\Geo\WKT\Exception\UnexpectedValueException
+     * @expectedExceptionMessage [Syntax Error] line 0, col 21: Error: Expected CrEOF\Geo\WKT\Lexer::T_CLOSE_PARENTHESIS, got "45.3" in value "LINESTRING(34.23 -87 45.3 -92)"
      */
-    public function testParsingMultiPointValueWithExtraParenthesis()
+    public function testParsingLineStringValueMissingComma()
     {
-        $value  = 'MULTIPOINT((0 0,10 0,10 10,0 10))';
+        $value  = 'LINESTRING(34.23 -87 45.3 -92)';
+        $parser = new Parser($value);
+
+        $parser->parse();
+    }
+
+    /**
+     * @expectedException        \CrEOF\Geo\WKT\Exception\UnexpectedValueException
+     * @expectedExceptionMessage [Syntax Error] line 0, col 26: Error: Expected CrEOF\Geo\WKT\Lexer::T_INTEGER, got ")" in value "LINESTRING(34.23 -87, 45.3)"
+     */
+    public function testParsingLineStringValueMissingCoordinate()
+    {
+        $value  = 'LINESTRING(34.23 -87, 45.3)';
         $parser = new Parser($value);
 
         $parser->parse();
